IJPL-160390 Mark frontend progresses as suspendable
Right now, the only way to mark progress as suspendable is to call `ProgressSuspender.markSuspendable`
We need to call this method on the frontend if received progress is marked as suspendable.

We also need to track changes in `ProgressSuspender` to actually pause the task on the backend.
And we need to subscribe to `task.statuses` to display the current status in the UI.

GitOrigin-RevId: 341f2dd93ae09731a738d9559d9ec68f11c9a89c