[gradle] IDEA-363330 prevent NPE in SyntheticVersionCatalogAccessor: use a model of exact catalog
In the accessor, use GradleVersionCatalogModel (a model for a single catalog) instead of GradleVersionCatalogsModel (a model for multiple catalogs). The last one also gives access to libraries, plugins, bundles and versions - but the result is nullable. Accessing them with `!!` looks suspicious to me. Probably, it was a cause of NPE while <init> of the accessor. Anyway, the usage of GradleVersionCatalogModel looks better to me.

GitOrigin-RevId: b659570aa14e89654ece34d81f6771706760caf2