[gradle][groovy] IDEA-357902 use GrLightMethodBuilder to create a method with Closure instead of LightMethodBuilder
For methods having a last parameter of Action type, GradleActionToClosureMemberContributor generates a decorating method with Closure instead of Action.

Before this commit, LightMethodBuilder was used for creating the method. I replaced its usage with GrLightMethodBuilder because looks more related to Groovy and because I need to add later some functionality not available in LightMethodBuilder (marking a created method as deprecated).

GitOrigin-RevId: d1959e9cae316bb14ac03f7b352199c1920b2ef9