IJPL-3020: simplify PerProjectIndexingQueue (per-provider sinks are not needed)
Sinks were initially introduced to reduce contention on lock that protects shared files queue. For some time already sinks work as a light-weight adapter for the shared queue (meaning that locks is used). We've never seen this lock in profiles since then.

There is a functional change however: cancelAllTasksAndWait method. It is not needed anymore: scanning should be cancelled through scanning executor (and it is cancelled through executor indeed for quite some time already).

GitOrigin-RevId: 3677aafb40e6eb47aa5b2f52001eef5cf88d0738