IJPL-160390 Fix ISE "Attempt to stop ProgressIndicator which is already stopped"
To mark an indicator as suspendable, 2 conditions have to be met:
- The indicator has to be running
- The indicator has to be the global indicator for the current context
See assertions in `com.intellij.openapi.progress.impl.ProgressSuspender.ProgressSuspender`

To meet these conditions, I call `markSuspendable` under `ProgressManager.getInstance().runProcess`, but `runProcess` stops the indicator when the execution is finished.

It is fine if the method is called before the indicator is shown in the status bar (before `PlatformTaskSupportKt.showIndicator` is called), because `showIndicator` controls when the indicator is started or stopped.

But now, this can happen mid-execution (we subscribe to `suspensionState` changes) which can lead to ISE "Attempt to stop ProgressIndicator which is already stopped" at `com/intellij/openapi/progress/impl/PlatformTaskSupport.kt:398`

The proper way here is to actually get rid of `ProgressIndicator` and subscribe to `TaskInfoEntity` directly in UI.
But that requires major changes, so for now I've added a hack:
`ProgressSuspender` has `attachToProgress` method which does pretty much the same as `markSuspendable`, but doesn't require conditions described above.
So, the `ProgressSuspender` is first created for a temp indicator and then attached to a real one.

We need this only for UI (it won't actually suspend anything), so it should be fine to mark indicator this way

GitOrigin-RevId: 40e3b53d86373836f369bc22ff2fd8c28e60a0d4