IJPL-160390 Fix finished indicator is never removed from the status bar
If `addProgress` is called for the already finished indicator, the progress bar is going to be stuck in the status bar forever (it won't be removed).
It happens because we expect to receive the `stop` signal when the indicator is finished, but it can only happen once.

So, we should check whether the indicator is still running before adding it to the status bar.
Also, it seems that we should override `finish` method instead of `stop`, as it is the method used by `InfoAndProgressPanel.MyInlineProgressIndicator.finish` to remove the progress.
On `stop` progress is not removed.

GitOrigin-RevId: f375325957f01a5d8b206f661a519515c95a6cdb