IJPL-173911 Fix menus and popups in Undocked mode
After the recent changes to the Undocked mode,
the tool window is now hidden before any action is performed
unless that action was initiated by a component inside the active
tool window.

Turns out this approach misses an important scenario when
the action is initiated from outside the tool window, but its
context component is inside. Popups are a perfect example:
a popup isn't a part of the tool window and yet its context
component may very well be inside.

This results to the context component being hidden along
with the tool window before the action can be performed.
Then the action manager refuses to perform it because
the context component isn't showing.

Fix by using the context component to determine which
tool window initiated the action, only falling back to
the input event component if there's no context component.

GitOrigin-RevId: 9b772436cdea6353d20a44733f2fd360363ec256