IJPL-172075 Fix custom external action icons
The problem with using getActionOrStub as the action supplier
to get the custom icon is that if the action isn't loaded yet,
then its stub will have a null icon. Therefore, customization
doesn't work.

Performance considerations: it shouldn't cause any serious
performance issues unless there are very many icon customizations,
which is normally not the case.

GitOrigin-RevId: 4fa084bdcaebe8e06dc26f4004a6eeac33dd36a1