[kotlin] Use `KaFunctionType.parameters` for properly rendering parameter names of `invoke` calls in inlay hints
Before that change, there was an issue with rendered parameter names when implicitly calling `invoke` on a function reference, as hints contained names of the `invoke` function itself (`p1`, `p2`, etc) and not the name of the actual function being invoked.
That is why rendering for implicit `invoke` calls was suspended.

^KTIJ-30438 fixed

GitOrigin-RevId: b46aa756f8d80b94c24e18ad838a7fbaa62990a6