[fleet] Fail gracefully when skiko is not downloaded at the right place + fix skiko path in some places
Before that commit, tests that are outside of our usual test framework (which would set `skiko.library.path` properly) would fail on a clean machine. If by chance the developer ran some `:fleet-build-project` task then it will work magically.
That commit remove the magic, it either work because the property is set by the framework, or fail gracefully if not and if skiko was not downloaded telling the developers how to resolve the issue.

Ideally, tests should be run by a single build tool (gutter run, or else should redirect to it). In that setup, we will wire prerequisites as part of the build tool instead of developers having to manually run something.

GitOrigin-RevId: dec85c07590160460d5b40dd135ed5d94fc2850c