[IJPL-174685] Do not continue processing if kotlin data node is null
the search for the dataNode uses a name comparison, and for the case of Android for example, the names won't match and we will end up finding nothing.
The function should not continue processing data if the dataNode is null because it will just use fallbacks instead of configuring the right way.
For the NAdroid case, we even have our own data service that actually passes the right name in this case.

closes https://github.com/JetBrains/intellij-community/pull/2898

(cherry picked from commit 08e60ab334209019dbca01bd226ada71a70012bd)

Bug: 364573852
GitOrigin-RevId: 96428ded562d3ff92e2de915e5afc4ae92384cd7
Change-Id: I990dff75998a92499c37c275d4700d0af6776df8