testFramework: prevent log records multiplication in console
When idea.test.logs.echo.debug.to.stdout is enabled and some
loggers have FINE or lower level allowed then some records might
get multiplicated in the output -- see comment in the code. This
was caused by too many handlers assigned to loggers and this change
reduces the amount of them.

GitOrigin-RevId: 903f52a3f1b01f02303a15b2c633892e13240a89