[remote dev] refactoring: introduce and use intellij.platform.backend module as a marker for modules not compatible with the frontend process
Before, we used intellij.platform.monolith module for that, but it looks strange when modules that can work at the backend have dependency on 'monolith'. And it may cause problems in the future if we start using separate installations for monolith and backend. So given that to fix IJPL-164312 we need to mark more modules as incompatible with frontend, it's better to use the proper marker.

GitOrigin-RevId: 329fdcd99a298e56baaf9dfdac75604ec512f84c