[terminal] IJPL-173788 Improve output scrolling strategy to take into account command block insets
There are additional insets between command blocks. Because of that simple 'scroll to bottom' strategy is not suitable anymore because the actual height of the output is now bigger, so scrolling to bottom can be excessive in some cases.
To leverage it, I had to implement more smart logic that considers the top line of the terminal screen and  cursor position for scroll offset calculation.

GitOrigin-RevId: bf22ff1d7f003e4f5e2492e9d038e56715d0ea4f