IJPL-173913 Don't hide a tool window if it's about to receive focus
Another hard-to-fix-reliably focus bug.

When one undocked tool window is activated, and another one
was previously activated, the focus doesn't immediately
go from one tool window to another. Instead, it's first transferred
to the editor before the new tool window is shown,
and then it's transferred to the new tool window.
This temporary focusing of the editor causes the newly
shown tool window to get hidden automatically,
which looks to the user as if it was never shown.

Fix by temporarily prohibiting auto-hide for a tool window
if it's about to receive focus. This fix relies on event processing
order: by the time invokeLater is processed, all focus events
should be already gone. In the current implementation
it's the case, so the fix works.

GitOrigin-RevId: 967a03251b4e2f188f841de806892de7b4a9fdd7