[kotlin] disable `JavaClassesInScriptDependenciesShortNameCache` for K2
According to the implementation of `JavaClassesInScriptDependenciesShortNameCache.getClassesByName`, it only provides classes from `JavaShortClassNameIndex` within the specified subscope.
`PsiShortNamesCacheImpl` performs a similar function but operates on the entire scope. Therefore, it seems redundant, at least for K2.

It affects performance in a bad way (see KTIJ-32683)

`JavaClassesInScriptDependenciesShortNameCache` is disabled only for K2, as in K1 it (may?) still be used in places where `PsiShortNamesCacheImpl` is not used.

^KTIJ-32683 fixed

GitOrigin-RevId: 1a9d24b41b4b372aa5a821a7cfcf20a894362967