[maven] IDEA-360900 MavenProjectReader remove reporting a problem when parent version is not present. Refactor.
Showing an error about missing parent version tag is already implemented in MavenParentMissedVersionInspection. No need to duplicate.

Also, fix estimating version using relative path when groupId and artefactId are not present in parent as well. I could not add tests for this because MavenProjectReader cannot execute tests on Maven 4 yet

GitOrigin-RevId: 990ee5062c4ef27a1c944dfe483b58371ec8e92b