IJPL-173321 Refactor ScopeEditorPanel: use selection paths directly
Using rows and then converting them to paths looks off by itself,
as we don't use the row indices anywhere. Moreover, in the default
Swing implementation, the getSelectionRows() function actually
takes the selection paths and then takes some extra effort to convert
them to rows, which we then convert back to paths. The performance
impact may or may not be significant, but given that we can both
get rid of it and simplify the code, it doesn't really matter.

GitOrigin-RevId: dab8e7df75e01eda099768bd9c4fbe2975e66d07