IJPL-173321 Check for node type in ScopeEditorPanel
Because the tree now uses an async model, sometimes
a node can be a LoadingNode. Such nodes should simply
be ignored by the selection updater, as normally the user
would wait for the real nodes to load and select them instead.
But in tests sometimes loading nodes are selected instead,
causing an exception for no reason.

GitOrigin-RevId: 52d3225d23f1347dd15f8648bddac2f87b789aab