build scripts: tests.cmd cannot find `JUnit5TeamCityRunnerForTestsOnClasspath` if executed locally
For local launch `intellij.build.use.compiled.classes` is always `true` because all classes are expected to be already prebuilt by the IDE `Build project` action, which isn't getting called if the script is launched from the terminal. And we don't want to mess with the classes compiled by the IDE, so we always assume that they are ready to use. The workaround is to explicitly specify `intellij.build.use.compiled.classes` and `intellij.build.incremental.compilation` in the tests.cmd script.

IJI-2080

GitOrigin-RevId: cec4e41a607165d5e1d8b44a06ed1fee2fd24702