[terminal] IJPL-175151 Make sure that line with cursor exists
There might be a problem when cursor is placed in a position that is not yet created in the text buffer (and the document). So, when we try to map the cursor position to the offset in the document - we get the wrong position. It was causing incorrect block start/end positions in Bash.

GitOrigin-RevId: 9ee55c3620abae70119b403d4651eade8344610f