[platform] always take OrderEnumerationHandler.shouldIncludeTestsFromDependentModulesToTestClasspath into account
Before, the option wasn't taken into account during processing dependencies recursively, because in that case source parts of dependencies are processed as 'ModuleSourceOrderEntry'. This didn't cause real problems because currently all plugins that return 'false' from 'shouldIncludeTestsFromDependentModulesToTestClasspath' also disable recursive processing via 'shouldProcessDependenciesRecursively'. However, it may cause problems in the future.

A test for this will be added in OrderEnumeratorHandlerRegistrationTest as a separate commit.

GitOrigin-RevId: de2f50d2f2cc5ec7700d47c35f62b4e11f842c44