KTIJ-26671 [kotlin] Avoid showing candidates with same fqName in auto-import popup in K2 Mode
Candidates with same fqName are always imported together,
so it doesn't make sense to show all of them to the user
and to propose to select one of them.

It should be OK to show only a single candidate in the
auto-import popup for such cases (like it's done in K1 Mode).

^KTIJ-26671 Fixed

GitOrigin-RevId: 2923d4bd98b1a5938fba54c28d9aa742069e42e5