[devkit] update runtime module repository builder to always generate descriptors for modules with production sources
Even if the name of the module satisfies 'isTestOnly' conditions, we still need to generate a descriptor for it if it has production sources. Recently, a new module 'intellij.rdct.tests.plugin' was added, which satisfies the condition, and at the same time the test which checks that no modules satisfying 'isTestOnly' condition have production sources was modified (see 39890e67352a1889). So the descriptor wasn't generated for the module, and build scripts fail it in FrontendModuleFilter.isModuleCompatibleWithFrontend.

GitOrigin-RevId: 2814a470f4825449c21adf5bc013c01bfd3f948b