IJPL-175085 Fix exceptions in `TaskInfoEntity` updates
Two exceptions are fixed here.
1. "Task 536870937 is not suspendable"

Task suspension is a mutable property for old progress indicators.
It can happen that the property has yet to be updated on this thread by the time `TaskManager` is called.
It'd be more reliable to check suspension property under change&shared blocks.

2. EntityDoesNotExistException

We used `tryWithEntities` to ensure that the entity exists when we access it.
However, turned out `tryWithEntities` works for `change` blocks but doesn't work for `shared` (in case of rebase)
Therefore, it could have happened that the entity was removed before the shared block and `trySetTaskStatus` will fail.

This exception, in turn, causes the kernel to be shut down because it happens in the rebase loop.

GitOrigin-RevId: 2aa4d9c11bb163d5f7733400ec2d8087e15bd12c