Refactor the `TaskStorage.updateTask` method to allow any updates (not state only)
`TaskStorage` provides the ability to update the task in the proper context (either local or shared) depending on whether the task was created on the backend of frontend side.

In `updateTask` we can also check for task's existence to ensure that updater won't be called for already deleted tasks (e.g. in case of rebase loop)

GitOrigin-RevId: 54153669876595ce2cb38950946f741adc3e7cf0