Fix potential races in task suspension updates
For old progresses we create the `BridgeTaskSuspender` which can change its suspension state from `NonSuspendable` to `Suspendable` and backwards.

Before we processed changes in `isSuspendable` and `taskInfo.statuses` in parallel, so it can easily happen that we try to pause a task (`TaskManager.pause`) before we actually made the task suspendable.
If it happens, the task will be suspended, but the UI won't display that.

To fix that, we should subscribe to status changes strictly after we've processed `isSuspendable`. And cancel the subscription once the task is no longer suspendable.

Also, when a task stops being suspendable, we need to ensure that it's not stuck in the ` Paused ` state by calling `TaskManager.resume`

GitOrigin-RevId: 8374892dcc32e9bf691a96e7cce8d0b9eea034d3