[RDCT] IJPL-171294: Fix storing window position and size for remote projects on frontend
This information is stored in `recentProjects.xml`. This was broken in `7ff4962e11bff286b82520862bb26a62a64b7907` by marking frontend projects as temporary, not saving any information in recent projects.
This commit fixes that by replacing this "marking temporary" API with an existing `hidden` info (which is used in Light Edit, for example), keeping info about recent projects and just not showing it in the recent projects list.

GitOrigin-RevId: f0452b05dd3ae287623f5c29361be870170e7e46