IJPL-175185 Repack the New popup when empty state changes
Because the empty text is relatively large, it doesn't look well
when the non-empty popup has only one or two lines.
Fix by repacking the popup when it becomes empty
or non-empty due to filtering.

To avoid affecting other popups, introduce an internal API
to only enable this feature for a specific popup.

GitOrigin-RevId: 3eca649ff30757d6ca4778a99d4e64f0d7c6b2ee