[ktor] IDEA-365732 Renamed the Ktor plugin package to com.intellij.ktor to align with our naming convention for packages
This change was necessary because ktor-gradle-tooling must share the same package name as the main Ktor plugin. Previously, due to the incorrect package name for Ktor, we had to use an inconvenient package name for ktor-gradle-tooling. To avoid this, we corrected the Ktor package as recommended by IntelliJProjectPackageNamesTest.root package name.

GitOrigin-RevId: 0d92edccc1bdbb896f8a4186404edc8c324e95de