IJPL-175257 Add `-XX:MaxHeapFreeRatio=45` to make heap shrink more willingly
The goal of this task is to tune JVM to improve the elasticity of the Java heap size,
 making it not only grow in the case of a memory-intensive workload (e.g., indexing),
 but also to shrink when the memory demand decreases.

We currently use G1 GC. The ability of G1 GC to grow and shrink the heap can be configured using `-XX:MinHeapFreeRatio` `-XX:MaxHeapFreeRatio` JVM options. The default values are 40/70 respectively.

After a (certain) garbage collection, the GC decides to grow or shrink the heap based on the percentage of free memory.
If the free memory percentage is lower than `MinHeapFreeRatio`, the heap is grown.
If the free memory percentage is higher than `MaxHeapFreeRatio`, the heap is shrunk.
A result of shrink or growth is that free memory percentage fits in between
(`MinHeapFreeRatio <= free_ration <= MaxHeapFreeRatio`).

The value of `-XX:MaxHeapFreeRatio=45` seems safe, none of the performance metrics degraded during the initial assessment.

GitOrigin-RevId: 4105f72a20fb5ef705073180be98fdf10e772f85