IJPL-175293 Fix actions from popups in unpinned tool windows
The problem is that when some action originates from a popup
and has its context component inside that popup,
then invoking auto-hide hides both the tool window and the popup.
This, in turn, makes the action subsystem think the action
is no longer relevant because its context component is no longer
visible.

Fix by checking that the context component is in the popup
and not auto-hiding anything in that case.
The check isn't perfect, as there seems to be no easy way,
so it relies on the fact that popup windows have the class
"javax.swing.Popup$HeavyweightWindow...something".

GitOrigin-RevId: 5412132e8cd2c94a2676f7bddc50141b05427812