Improve AboutPopupDescriptionProvider API
The EP now offers a way to provide different text to append to the
"extended text" that the dialog copies when the Copy button is clicked.
Extended descriptions can provide multiple lines of plain text. This is
useful e.g., to enumerate non-default flags in Android Studio's own flag
system (similar to the IJP Registry).

By default, this is the same as what the existing getDescription() API
returns, so the behaviour is not changed for existing EPs. The KDocs
have been updated and improved to better reflect the contract of the EP.

In addition, the AboutDialog has been changed to:
 * Use getExtendedDescription() to build the "extended text"
 * Strip HTML that may be produced by the EP implementers, as it's
   supposed to be plaintext only, as per Yan's comment on
   https://youtrack.jetbrains.com/issue/IJPL-1813
 * Ensures the non-extended description only provides one line of text,
   as per Yan's comment on IJPL-1813