IJPL-175215 Fix stuck progress in the status bar
Some progresses might get stuck in the status bar if the `TaskInfoEntity` is removed before `showIndicatorInUI` is finished.

`showIndicator` is wrapped into `tryWithEntities(task)` in `TaskInfoEntityCollector`.
`tryWithEntities` cancels running jobs if the task is deleted.

The problem was that `showIndicatorInUI` got canceled mid-execution due to task deletion.
In particular, during service retrieval (`ComponentManagerImpl.getService`) inside `ProcessBalloon.addIndicator`

Therefore, we don't call the `indicator.finish` method which is supposed to remove the indicator from the status bar.

GitOrigin-RevId: 585c548f00910252d2e24d15148043ae168f4e45