IJPL-175215 Cancel the indicator only if it was explicitly canceled from UI
The indicator might be removed from the status bar without being stopped if `finish()` method was called without `stop` or `cancel`.
This means that the indicator might still be running even if it's not shown.

We, however, were cancelling the indicator every time it's removed from the status bar if it is still running.

To understand that the execution was canceled by UI, we can catch `CancellationException` which is going to be thrown when the "cancel" button is pressed. And we should wait for `indicatorFinished` instead of `awaitCancellation()`, because the latter throws the exception in any scenario.

So, now we wait for `indicatorFinished` to complete. If it's completed without a cancellation exception, it means `indicator.finish(taskInfo)` was called and we don't need to do anything else.

If `CancellationException` happens, it means that `withBackgroundProgress` has been canceled and the only way to cancel it is from UI. In this case we should check whether the indicator is still running and cancel it.

GitOrigin-RevId: 15ee74862a5a1f381faaf6c5c214f37d7e47bfeb