[RDCT] IJPL-166131: Support product workspace files in Remote Dev
Project store doesn't use `SharedConfigFolderStreamProvider`, so all `workspace/<project-id>.xml` files are written to `per_process_config_*` folders for JetBrains Client. It prevents saving a lot of project workspace settings because per-process folders do not correspond to a certain project and are cleared on every project opening.
This commit moves writing of these files to the original config directory, which is safe, because for multiple client processes they will have different opened projects, so they will not write to the same file on disk.

GitOrigin-RevId: af15b84236d5582e16f40dc89bbf004a11cff413