IJPL-165130 [reworked terminal] Introduce SendShortcutToTerminalAction
We have a list of actions that are allowed to be executed
in the terminal tool window.
When a key event is received, there are different possibilities:

1. It corresponds to no actions at all.
2. It corresponds to some actions,
but none of them are on the allowed list.
3. It corresponds to some of the allowed actions,
and at least one of those is enabled.
4. It corresponds to some of the allowed actions,
but none of them are enabled.
At the same time, there's another platform action
with the same shortcut that is enabled.
5. It corresponds to some of the allowed actions,
but neither they nor any other corresponding
actions are enabled.

We have three ways to process a key event:
A. Before the action system (using a dispatcher
or a toolkit listener).
B. Inside the action system, by registering an action.
C. After the action system, by installing a key event
listener on the component.

Way A catches all events, but we don't know which
actions are enabled at that moment.
Way B only works if we have an action
corresponding to the event being processed.
Way C only works for events that correspond
to no enabled actions.

Cases 1 and 2 are handled using way A:
we handle all key events that don't have a corresponding
action on the allowed list.
Case 3 is handled by the action system.
The corresponding action is performed as usual.
Case 5 used to be handled using the way C.
But the way C couldn't handle case 4.
The action system would then perform the action
we don't want to be performed.

Fix by using way B to handle cases 4 and 5.
Register an action (SendShortcutToTerminalAction)
that handles all shortcuts of the all allowed actions.
This action is only enabled if all allowed actions are disabled.
And it sends the shortcut directly to the terminal.
It's also promoted, so the platform considers it before
other actions.

We could continue handling case 5 the old way,
but now that we have an action that catches all
shortcuts anyway, the key event handler
is simply not invoked on such key events.
So we drop the key listener.

GitOrigin-RevId: 641442bc75a3f423b943dda940c7aede0a174083