PY-78299 Fix flaky "Unused local" inspection with runForWholeFile="true"
The problem occurs when there are nested functions referring to variables
in the outer scope. The inspection implementation requires all nested scopes
to be visited to correctly detect unused local definitions, but the platform
may skip inner function when their enclosing function changes.

Unlike in PY-77891, the problem here is not the order of `visitPy*Element` calls
in the inspection visitor, but the fact that for some elements it's not invoked
at all.

Using runForWholeFile="true" forces full tree traversal and remedies the problem.

GitOrigin-RevId: 59ecba5ca204d51bba3e3b1b73c03f70bcbe231c