IJPL-175215 Check that `ProjectEntity` still exists when we create a new task
`TaskInfoEntity` is created under the `change+shared` block. It might happen that the `ProjectEntity` is already removed by the time we enter the block to create a new entity.

We should double-check that the project still exists inside `createTaskInfoEntity` and if it doesn't, we can throw `CancellationException` to cancel the execution

GitOrigin-RevId: 4a75fc4f6ecdc97bd1dd275d5a1d9c31f588e70c