IJPL-175215 Create a task info entity in parallel to task execution
To avoid any delays in execution or context switching, which caused some tests to fail (e.g. `SettingsSyncPluginManagerTest` and `QodanaHighlightedReportServiceTest`) we should create and update `TaskInfoEntity` in parallel.

It's not really necessary to wait for task creation before starting the execution.
The task is only required for UI, and we don't wait for progress to be shown anyway.

Moreover, creating a task in the parallel will allow us to avoid stopping a task from execution if something goes wrong with `TaskInfoEntity` creation.

GitOrigin-RevId: e95fd01f16ff78091baaa16074ba68fb5fbb4813