IJPL-175185 Fix empty / non-empty popup repacking
Packing when the popup becomes empty is fine,
but packing when it becomes non-empty is not
because it depends on the first non-empty state.
If the speed search field is modified in such a way
that a previously empty popup now has only
a few lines, it'll become small and remain at that size.

Fix by not using pack() to bring back the non-empty size.
Instead, save the last size the popup had when it was
non-empty and reuse that size when it becomes non-empty
again.

GitOrigin-RevId: 8c7bc99d515b648c5089c4ea7b0852550b6285d8