[JEWEL-742] Load markdown images using Coil3
It supports every image as an inline node;
Using built-in coroutine library and ktor3 from the platform;
Added SVG support using a coil dependency.

Moved images into an extension so it can be loaded
on demand without pushin coil3 dependencies to everyone.
Add Coil3ImagesRendererExtension to rendering extensions to
render images.

Also expose `imageLoader` as a parameter to give fine
controls to apps wanting to pass in their global loader.

GitOrigin-RevId: f7b51b8239622c39297aeb8dcfd8fad3435111ec