Move thenIf modifier to foundation
This helps with places like the incoming SimpleTableLayout that need it
in `foundation`, but don't have access to it because it's in `ui`.

It also makes semantic sense, since the other, similar, modifiers we
created all live in foundation; thenIf has no "ui" specificity to it.

This also marks the "old" modifier as scheduled for removal in 252.