IJPL-165667 [reworked terminal] Remove SendShortcutToTerminalAction from global actions
Using it both as a global action and using registerCustomShortcutSet
produces a warning. It makes sense, too, as it's a local action
only available when the terminal is focused, so it has no business
in the global action system.

Split the action and promoter to make the promoter more lightweight.
Rewrite the promoter to detect the action using a client property
to deal with wrappers / delegates.

GitOrigin-RevId: 0da3cf704751f2f856ee2d46b335a38ad2f02b30