[java] optimization: introduce JavaPsiFacade::hasClass method and use it in JavaLibraryUtil (IJPL-166229)
If there are multiple classes with the provided name in the scope, 'findClass' and 'findClasses' sort them according to their positions in the classpath. This requires building the OrderEntryGraph, which may take considerable time in a big project (see IJPL-930). New methods 'hasClass' are added to JavaPsiFacade, PsiElementFinder, JavaFileManager. Their implementations don't need to sort the results and therefore don't require building OrderEntryGraph. JavaLibraryUtil is migrated to use the new method.

GitOrigin-RevId: 0b4b05f54a5234e4e585c293bd9bd4a1957a4d70