IJPL-165667 [reworked terminal] Fix Terminal.Next/PrevSplitter
The logic in
TerminalSessionContextMenuActionBase.findContextTerminal
only worked for non-split terminals. For split terminals
it always returns the first terminal in the splitter,
so the next action always went to the second terminal
and the previous action went to the last terminal.

Fix by directly providing the focused widget through
its wrapper panel into the data context.

The classic terminal isn't affected, as it doesn't use
findContextTerminal, it uses a listener in
TerminalToolWindowManager.setupTerminalWidget
to invoke the action instead.

GitOrigin-RevId: 26cf3c8a0cd1cbb8239d7d6fa93311c4ac99686a