IJPL-165667 Fix the ShowContent action
A very old bug caused by the wrong order of arguments:
the action would be enabled if the tool window is descending
from the context component, which doesn't make sense
and is almost never the case. It should be the other way around.

Most tool windows aren't affected by this because
they register their own copy of this action in ToolWindowContentUi.
However, in the terminal this copy doesn't work because
it doesn't pass the allowed-list check. Only actions properly
registered in the action managers with IDs on the allowed
list work in the terminal.

GitOrigin-RevId: 1b4cc5d976fa8ce6c9c20eaa7921cec30cc74ac1