KTIJ-32821 [kotlin] Introduce `UnresolvedReferenceDiagnosticsRestoration` utilities
They would allow passing actual diagnostics from `KotlinDiagnosticHighlightVisitor` to `ImportQuickFixProvider`,
instead of relying solely on PSI elements.

Unfortunately, the diagnostics cannot be safely passed between Analysis Sessions
as is - they can become invalid.
To overcome this, we collect the diagnostics for the target elements again.

If KT-74573 is ever fixed, this approach can be changed to store
pointers and restore them in `ImportQuickFixProvider`.

GitOrigin-RevId: 9d686eacc9c17cc0dcbf594115c6e6020c5ee9bc