[plugin model] refactoring: extract ModulesWithDependencies class and use it instead of ModuleGraph
Actually, we need to build the module graph only to sort modules and their dependencies. There is no need to sort the graph itself or expose it to other classes. This change simplifies logic and allows having edges in the graph which don't correspond to dependencies between modules (this is needed for IJPL-175069).

GitOrigin-RevId: 197ff017499e462ad655764f1f5f10606d354782