IJPL-175596 Fix progresses might not be shown at the start
`TaskInfoEntityCollector` is a service that gathers active tasks and sent them to `IdeStatusBart`
Because the service was preloaded, it might have happened that the status bar was not created at the moment a task appears.

It was especially noticeable in the RD scenario because there are already some tasks stored in DB when the frontend is opened.

`IdeStatusBar` should subscribe to tasks by itself when it's created, but it's also created too early when the project is yet to initialize.

The best place for the subscription seems to be `ProjectFrameHelper.setProject` - we know that both the project and the status bar are initialized and the task collection process can be started.

GitOrigin-RevId: 920953988548c0fd8f45184dd3eb5cccce3b95ac