IJPL-165667 [reworked terminal] Implement some Escape handlers
To be able to handle Escape properly without suppress / promote
hacks, let's introduce an EP for this.

To allow easy prioritizing of escape handlers,
for now let's go with a simple solution: just define
an integral order for every extension.
Then they can be sorted and considered in that order.

At this point only the Cancel Selection and Focus Editor
handlers are implemented.

GitOrigin-RevId: 33086f5630bd413d34eca3581c5fac7684fb1644