IJPL-175730 Fix freeze when opening a new project from VCS when another one is already open
The reason why this happened is that we convert old `ProgressIndicator` to a new API (`withBackgroundProgress`) using coroutineScope attached to ide status bar.

This scope is canceled when the connected project is closed and `CancellationException` is thrown.
We, in turn, cancel the indicator when `CancellationException` is received.
The original of handling `CancellationException` and canceling the indicator after, was that a user can cancel the progress via UI.

However, we didn't expect the situation where the scope is canceled from the outside (on project closure).
In this case we should let the indicator continue running.

To achieve that, I've created an app-level service which runs `withBackgroundProgress` under its coroutine scope, which won't cancel on project closure.

GitOrigin-RevId: 6dd33b7a3069bbea579f84244729a1beacf9eb70