IJPL-175730 Fix "Project is null for task" error
IdeStatusBar calls `registerCloneTasks` on creation, which is supposed to add all running tasks (and subscribe to new).
However, `project` is yet to be initialized at this point, so we get the error "Project is null for task" in `BridgeTaskSupport` and the task is not added to the status bar.

`withBackgroundProgress` requires a project to be not-null, so we need to wait for it to be initialized before we add any progresses.

GitOrigin-RevId: afd2fdb2be87f896fe0a14ba99ff6e90365ba58e