IJPL-165667 [reworked terminal] Fix caret exceptions
If a document update makes it smaller, for example,
when invoking the Clear Buffer action, the cursor isn't updated
immediately. In the meantime, it may be accessed
and that throws exceptions, as the state is incorrect.

Ideally, this should be fixed by redesigning the model in such a way
that such state updates are atomic, but for now a simple
consistency check will do. If the cursor is outside the document
after an update, make it fit into the document immediately,
and update it to the correct position later.

GitOrigin-RevId: 85f637fbd480cb62a1b42312b22fbbbd9396328b