[RDCT] IJPL-166131: Fix error from ClientAwareComponentManager
There is a bug in `ClientAwareComponentManager`, if you call `project.serviceOrNull()` at the project's pre-init stage, from `postGetService` it will call `super.doGetService` for sessions manager, which will result in error "ClientSessionsManager is registered as application service, but requested as a project one".
This commit fixes that by avoiding calling `serviceOrNull` at the pre-init stage of a project loading.

GitOrigin-RevId: e7e08195d4ddc9f93e0f74b513487ee8876c9f5a