[plugin model] support dependencies between required content modules in a plugin (IJPL-175069)
Before, edges in ModuleGraph were also used to configure dependencies in ClassLoaderConfigurator. This doesn't work well in some cases. For example, because we need to process all required modules before the main plugin module to correctly determine whether the plugin can be loaded, but don't want to have dependencies from the main plugin module to all required content modules (this may lead to circular dependencies). So now when ModulesWithDependencies is computed, additional edges are stored in a separate map and used only to build the graph.

GitOrigin-RevId: 84f47dbeeb88fa6c0953bacd83f8d8eb59587dbd