IJPL-165667 [reworked terminal] SendShortcutToTerminalAction: compute shortcuts
Because shortcuts can change,
it's not a good idea to compute them once.
Computing them every single time
is OK here performance-wise, because registering
only happens on focus change,
that is, not that often.
And it's guaranteed that there will
be a focus change if the user goes
to the settings to change some shortcuts,
so there's little risk having outdated shortcuts there.

GitOrigin-RevId: 073afe42fb564146560ea2c576800d2b47abc159