[debugger-rd] IJPL-160146: Introduce frontend mark object action
Also, it required to change how FrontendXValues are getting disposed. Before the change such xValues were disposed when popup is closed, and now they will be disposed only when evaluator is changed (when another stack frame is selected). It was needed, since Mark Object action opens another dialog, so XValues got disposed and we wasn't able to mark such values on backend.

GitOrigin-RevId: a044c7d4203e8f335428a76fb7d381a3a0ce8ba2