[build scripts] download JARs from all project-level Maven libraries before starting build (IJPL-175909)
Before, JARs were downloaded only for libraries used in modules. This contradicted the behavior of the IDE and sometimes caused problems on TeamCity. A special 'fleet.compiler.plugins' module was introduced as a workaround to refer to libraries which JARs were used as Kotlin compiler plugins but weren't added as dependencies directly.

GitOrigin-RevId: 078b8741282a7ede2d3ab10b617bffaf24dd65bc