IJPL-175872 drop the ability to search for all references or declarations of a symbol
The generic ability does not make sense because each feature (Find usages, Rename, etc.) is expected to implement its
own search. We've already had `ReferencesSearch` and tried to implement different features on top of it, and it turned
out that it's too generic. For instance, one needs to filter out some of its results during the Rename, but include them
in Find Usages. Instead, we require language support to implement different searchers for different features. If the
code is actually share-able, it's up to the language plugin implementation to share it.

GitOrigin-RevId: 65f719838645d780effcbe80a1aafaf93e756ae0