IJPL-165667 [reworked terminal] Fix ShowContent in the terminal
The previous fix (the wrong isDescendingFrom order)
worked fine for shortcuts, but in case of the context menu,
it doesn't work because there's no active window.
This is because the context menu itself steals focus from
the window.

Fix by getting the tool window from the data context
as a fallback. Perhaps it should be just the default implementation:
it's not clear why all that guesswork is needed at all.
But since it's old code, let's stay on the safe side for now.

GitOrigin-RevId: a32c0e60272ca87487bb2738a22a335dd5b6cfeb